+++ modules/processing/targetinfo.py
@@ -6,8 +6,12 @@ from __future__ import absolute_import
 import os.path
 
 from lib.cuckoo.common.abstracts import Processing
+from lib.cuckoo.common.config import Config
+from lib.cuckoo.common.integrations.file_extra_info import static_file_info
+from lib.cuckoo.common.integrations.parse_url import HAVE_WHOIS, URL
 from lib.cuckoo.common.objects import File
-from lib.cuckoo.common.utils import is_text_file
+
+processing_conf = Config("processing")
 
 
 class TargetInfo(Processing):
@@ -23,18 +27,24 @@ class TargetInfo(Processing):
         # We have to deal with file or URL targets.
         if self.task["category"] in ("file", "static"):
             target_info["file"] = {}
-            # Let's try to get as much information as possible, i.e., the
-            # filename if the file is not available anymore.
+            # Let's try to get as much information as possible, i.e., the filename if the file is not available anymore.
             if os.path.exists(self.file_path):
                 target_info["file"], pefile_object = File(self.file_path).get_all()
                 if pefile_object:
                     self.results.setdefault("pefiles", {}).setdefault(target_info["file"]["sha256"], pefile_object)
 
-                with open(self.file_path, "rb") as f:
-                    is_text_file(target_info["file"], self.analysis_path, 8192, f.read())
+                static_file_info(
+                    target_info["file"],
+                    self.file_path,
+                    self.task["id"],
+                    self.task.get("package", ""),
+                    self.task.get("options", ""),
+                    self.analysis_path,
+                )
 
             target_info["file"]["name"] = File(self.task["target"]).get_name()
         elif self.task["category"] == "url":
             target_info["url"] = self.task["target"]
-
+            if HAVE_WHOIS and processing_conf.static.whois.enabled:
+                self.results["url"] = URL(self.task["target"]).run()
         return target_info